Server integration done. Completes most of the data node work minus replication. End to end works End to end tests Coordinator code to support all API. This is temporary used for testing.
TODO
1. Lots more unit tests
2. Integration tests
3. Store Manager / Crash Recovery
4. Benchmarking

+++ ambry-messageformat/src/test/java/com.github.ambry.messageformat/MessageFormatInputStreamTest.java
@@ -110,7 +110,7 @@ public class MessageFormatInputStreamTest {
     Assert.assertEquals(streamData.readShort(), 1);
     Assert.assertEquals(streamData.readLong(), 2000);
     for (int i = 0; i < 2000; i++) {
-      Assert.assertEquals(streamData.read(), data[i]);
+      Assert.assertEquals((byte)streamData.read(), data[i]);
     }
     long crcVal = crcstream.getValue();
     Assert.assertEquals(crcVal, streamData.readLong());
